refactor(slider): move slider to new file
Moves slider from index.tsx to Slider.tsx
