+++ packages/core/src/Slider/slider.test.tsx
